+++ lib/compact_tweet_view.dart
@@ -59,6 +59,9 @@ class CompactTweetView extends StatelessWidget {
   /// If set to false a image placeholder will he shown and a video will be played in a new page.
   final bool useVideoPlayer;
 
+  /// If the Tweet contains a video then an initial volume can be specified with a value between 0.0 and 1.0.
+  final double videoPlayerInitialVolume;
+
   /// Function used when you want a custom image tapped callback
   final OnTapImage onTapImage;
 
@@ -79,6 +82,7 @@ class CompactTweetView extends StatelessWidget {
       this.quoteBackgroundColor,
       this.backgroundColor,
       this.useVideoPlayer,
+      this.videoPlayerInitialVolume,
       this.onTapImage,
       this.createdDateDisplayFormat}); //  TweetView(this.tweetVM);
 
@@ -97,6 +101,7 @@ class CompactTweetView extends StatelessWidget {
       this.quoteBackgroundColor = Colors.white,
       this.backgroundColor = Colors.white,
       this.useVideoPlayer = true,
+      this.videoPlayerInitialVolume = 0.0,
       this.onTapImage,
       this.createdDateDisplayFormat})
       : _tweetVM = TweetVM.fromApiModel(tweet, createdDateDisplayFormat);
@@ -142,6 +147,7 @@ class CompactTweetView extends StatelessWidget {
                           _tweetVM,
                           ViewMode.compact,
                           useVideoPlayer: useVideoPlayer,
+                          videoPlayerInitialVolume: videoPlayerInitialVolume,
                           onTapImage: onTapImage,
                         ),
                         GestureDetector(
@@ -152,6 +158,7 @@ class CompactTweetView extends StatelessWidget {
                             _tweetVM,
                             textStyle: textStyle,
                             clickableTextStyle: clickableTextStyle,
+                            padding: const EdgeInsets.only(top: 0.0),
                           ),
                         ),
                         (_tweetVM.quotedTweet != null)