+++ lib/embedded_tweet_view.dart
@@ -28,6 +28,9 @@ class EmbeddedTweetView extends StatelessWidget {
   /// If set to false a image placeholder will he shown and a video will be played in a new page.
   final bool useVideoPlayer;
 
+  /// If the Tweet contains a video then an initial volume can be specified with a value between 0.0 and 1.0.
+  final double videoPlayerInitialVolume;
+
   /// Function used when you want a custom image tapped callback
   final OnTapImage onTapImage;
 
@@ -39,6 +42,7 @@ class EmbeddedTweetView extends StatelessWidget {
     this.backgroundColor,
     this.darkMode,
     this.useVideoPlayer,
+    this.videoPlayerInitialVolume,
     this.onTapImage,
     this.createdDateDisplayFormat,
   }); //  TweetView(this.tweetVM);
@@ -47,6 +51,7 @@ class EmbeddedTweetView extends StatelessWidget {
       {this.backgroundColor = Colors.white,
       this.darkMode = false,
       this.useVideoPlayer = true,
+      this.videoPlayerInitialVolume = 0.0,
       this.onTapImage,
       this.createdDateDisplayFormat})
       : _tweetVM = TweetVM.fromApiModel(tweet, createdDateDisplayFormat);
@@ -84,30 +89,28 @@ class EmbeddedTweetView extends StatelessWidget {
                                 Padding(
                                   child: RetweetInformation(
                                     _tweetVM,
-                                    retweetInformationStyle:
-                                        defaultEmbeddedRetweetInformationStyle,
+                                    retweetInformationStyle: defaultEmbeddedRetweetInformationStyle,
                                   ),
                                   padding: EdgeInsets.only(left: 28),
                                 ),
                                 Row(
                                   children: <Widget>[
                                     ProfileImage(tweetVM: _tweetVM),
-                                    Padding(
-                                      padding: const EdgeInsets.only(left: 8.0),
-                                      child: Byline(
-                                        _tweetVM,
-                                        ViewMode.standard,
-                                        userNameStyle: TextStyle(
-                                          color: (darkMode)
-                                              ? Colors.white
-                                              : Colors.black,
-                                          fontSize: 16.0,
-                                          fontFamily: 'Roboto',
-                                          fontWeight: FontWeight.w700,
+                                    Expanded(
+                                      child: Padding(
+                                        padding: const EdgeInsets.only(left: 8.0),
+                                        child: Byline(
+                                          _tweetVM,
+                                          ViewMode.standard,
+                                          userNameStyle: TextStyle(
+                                            color: (darkMode) ? Colors.white : Colors.black,
+                                            fontSize: 16.0,
+                                            fontFamily: 'Roboto',
+                                            fontWeight: FontWeight.w700,
+                                          ),
+                                          showDate: false,
+                                          userScreenNameStyle: defaultEmbeddedUserNameStyle,
                                         ),
-                                        showDate: false,
-                                        userScreenNameStyle:
-                                            defaultEmbeddedUserNameStyle,
                                       ),
                                     ),
                                   ],
@@ -127,16 +130,11 @@ class EmbeddedTweetView extends StatelessWidget {
                     onTap: () {
                       openUrl(_tweetVM.tweetLink);
                     },
-                    child: Padding(
-                      padding: const EdgeInsets.symmetric(
-                          horizontal: 8.0, vertical: 15.0),
-                      child: TweetText(
-                        _tweetVM,
-                        textStyle: (darkMode)
-                            ? defaultEmbeddedDarkTextStyle
-                            : defaultEmbeddedTextStyle,
-                        clickableTextStyle: defaultEmbeddedClickableTextStyle,
-                      ),
+                    child: TweetText(
+                      _tweetVM,
+                      textStyle: (darkMode) ? defaultEmbeddedDarkTextStyle : defaultEmbeddedTextStyle,
+                      clickableTextStyle: defaultEmbeddedClickableTextStyle,
+                      padding: const EdgeInsets.symmetric(horizontal: 8.0, vertical: 15.0),
                     ),
                   ),
                   (_tweetVM.quotedTweet != null)
@@ -144,15 +142,10 @@ class EmbeddedTweetView extends StatelessWidget {
                           padding: EdgeInsets.only(top: 8.0, bottom: 10),
                           child: QuoteTweetViewEmbed.fromTweet(
                             _tweetVM.quotedTweet,
-                            textStyle: TextStyle(
-                                color:
-                                    (darkMode) ? Colors.white : Colors.black),
+                            textStyle: TextStyle(color: (darkMode) ? Colors.white : Colors.black),
                             clickableTextStyle: defaultQuoteClickableTextStyle,
-                            userNameStyle: (darkMode)
-                                ? defaultEmbeddedDarkQuoteUserNameStyle
-                                : defaultQuoteUserNameStyle,
-                            userScreenNameStyle:
-                                defaultQuoteUserScreenNameStyle,
+                            userNameStyle: (darkMode) ? defaultEmbeddedDarkQuoteUserNameStyle : defaultQuoteUserNameStyle,
+                            userScreenNameStyle: defaultQuoteUserScreenNameStyle,
                             backgroundColor: null,
                             borderColor: null,
                             onTapImage: onTapImage,
@@ -169,6 +162,7 @@ class EmbeddedTweetView extends StatelessWidget {
               _tweetVM,
               ViewMode.standard,
               useVideoPlayer: useVideoPlayer,
+              videoPlayerInitialVolume: videoPlayerInitialVolume,
               onTapImage: onTapImage,
             ),
           ),
@@ -183,18 +177,10 @@ class EmbeddedTweetView extends StatelessWidget {
                 ),
                 Container(
                     margin: EdgeInsets.only(left: 6),
-                    child: Text(_tweetVM.favoriteCount.toString(),
-                        style: TextStyle(
-                            color: (darkMode)
-                                ? Colors.grey[400]
-                                : Colors.grey[600]))),
+                    child: Text(_tweetVM.favoriteCount.toString(), style: TextStyle(color: (darkMode) ? Colors.grey[400] : Colors.grey[600]))),
                 Container(
                     margin: EdgeInsets.only(left: 16),
-                    child: Text(_tweetVM.createdAt,
-                        style: TextStyle(
-                            color: (darkMode)
-                                ? Colors.grey[400]
-                                : Colors.grey[600])))
+                    child: Text(_tweetVM.createdAt, style: TextStyle(color: (darkMode) ? Colors.grey[400] : Colors.grey[600])))
               ],
             ),
           ),
@@ -215,14 +201,17 @@ class EmbeddedTweetView extends StatelessWidget {
                     Icons.person_outline,
                     color: (darkMode) ? Colors.blue[100] : Colors.blue[700],
                   ),
-                  Padding(
-                    padding: const EdgeInsets.only(left: 5),
-                    child: Text(
-                      "${_tweetVM.userName}'s other tweets",
-                      style: TextStyle(
-                          color:
-                              (darkMode) ? Colors.blue[100] : Colors.blue[800],
-                          fontWeight: FontWeight.w400),
+                  Expanded(
+                    child: Padding(
+                      padding: const EdgeInsets.only(left: 5),
+                      child: Text(
+                        "${_tweetVM.userName}'s other tweets",
+                        style: TextStyle(color: (darkMode) ? Colors.blue[100] : Colors.blue[800], fontWeight: FontWeight.w400),
+                        maxLines: 1,
+                        overflow: TextOverflow.fade,
+                        softWrap: false,
+                        textAlign: TextAlign.start,
+                      ),
                     ),
                   )
                 ],