+++ lib/tweet_view.dart
@@ -58,6 +58,9 @@ class TweetView extends StatelessWidget {
   /// If set to false a image placeholder will he shown and a video will be played in a new page.
   final bool useVideoPlayer;
 
+  /// If the Tweet contains a video then an initial volume can be specified with a value between 0.0 and 1.0.
+  final double videoPlayerInitialVolume;
+
   /// Function used when you want a custom image tapped callback
   final OnTapImage onTapImage;
 
@@ -79,6 +82,7 @@ class TweetView extends StatelessWidget {
     this.quoteBackgroundColor,
     this.backgroundColor,
     this.useVideoPlayer,
+    this.videoPlayerInitialVolume,
     this.onTapImage,
     this.createdDateDisplayFormat,
   }); //  TweetView(this.tweetVM);
@@ -97,6 +101,7 @@ class TweetView extends StatelessWidget {
       this.quoteBackgroundColor = Colors.white,
       this.backgroundColor = Colors.white,
       this.useVideoPlayer = true,
+      this.videoPlayerInitialVolume = 0.0,
       this.onTapImage,
       this.createdDateDisplayFormat})
       : _tweetVM = TweetVM.fromApiModel(tweet, createdDateDisplayFormat);
@@ -111,6 +116,7 @@ class TweetView extends StatelessWidget {
             _tweetVM,
             ViewMode.standard,
             useVideoPlayer: useVideoPlayer,
+            videoPlayerInitialVolume: videoPlayerInitialVolume,
             onTapImage: onTapImage,
           ),
           GestureDetector(
@@ -139,13 +145,15 @@ class TweetView extends StatelessWidget {
                               Row(
                                 children: <Widget>[
                                   ProfileImage(tweetVM: _tweetVM),
-                                  Padding(
-                                    padding: const EdgeInsets.only(left: 8.0),
-                                    child: Byline(
-                                      _tweetVM,
-                                      ViewMode.standard,
-                                      userNameStyle: userNameStyle,
-                                      userScreenNameStyle: userScreenNameStyle,
+                                  Expanded(
+                                    child: Padding(
+                                      padding: const EdgeInsets.only(left: 8.0),
+                                      child: Byline(
+                                        _tweetVM,
+                                        ViewMode.standard,
+                                        userNameStyle: userNameStyle,
+                                        userScreenNameStyle: userScreenNameStyle,
+                                      ),
                                     ),
                                   ),
                                 ],
@@ -165,13 +173,11 @@ class TweetView extends StatelessWidget {
                   onTap: () {
                     openUrl(_tweetVM.tweetLink);
                   },
-                  child: Padding(
+                  child: TweetText(
+                    _tweetVM,
+                    textStyle: textStyle,
+                    clickableTextStyle: clickableTextStyle,
                     padding: const EdgeInsets.symmetric(horizontal: 8.0),
-                    child: TweetText(
-                      _tweetVM,
-                      textStyle: textStyle,
-                      clickableTextStyle: clickableTextStyle,
-                    ),
                   ),
                 ),
                 (_tweetVM.quotedTweet != null)